Reorganisation des composants dans leurs propres dossiers
