+++ src/sdg/_components.js
@@ -1,9 +1,9 @@
-export * from './components/notice.svelte'
+export * from './components/Notice/notice.svelte'
 export * from './components/PivHeader/pivHeader.svelte'
-export * from './components/pivFooter.svelte'
-export * from './components/alert.svelte'
-export * from './components/toTop.svelte'
-export * from './components/externalLink.svelte'
+export * from './components/PivFooter/pivFooter.svelte'
+export * from './components/Alert/alert.svelte'
+export * from './components/ToTop/toTop.svelte'
+export * from './components/ExternalLink/externalLink.svelte'
 export * from './components/SearchBar/searchBar.svelte'
 export * from './components/SearchInput/SearchInput.svelte'
-export * from './components/Icon.svelte'
+export * from './components/Icon/Icon.svelte'
