+++ src/sdg/components/ExternalLink/externalLink.svelte
@@ -7,7 +7,7 @@
 }}"/>
 
 <script>
-import {Utils} from "./utils";
+import {Utils} from "../utils";
 
 const {
     externalIconAlt = Utils.getPageLanguage() === 'fr'
