+++ src/sdg/components/Icon/Icon.svelte
